fix: RoundRecordRepository 집계 함수 반환 타입 Long으로 변경 #90
- JPQL의 COUNT, SUM 함수 반환값이 Long임에 따라 Repository 메서드 반환 타입 수정
- Service 및 DTO에서 불필요한 형변환 제거하고 Long 타입으로 통일

+++ src/main/java/com/hrr/backend/domain/round/repository/RoundRecordRepository.java
@@ -45,9 +45,9 @@ public interface RoundRecordRepository extends JpaRepository<RoundRecord, Long> 
 
     // 내가 참여한 라운드 개수 (몇 라운드째 참여 중인지)
     @Query("SELECT COUNT(r) FROM RoundRecord r WHERE r.userChallenge.id = :userChallengeId")
-    Integer countByUserChallengeId(@Param("userChallengeId") Long userChallengeId);
+    Long countByUserChallengeId(@Param("userChallengeId") Long userChallengeId);
 
     // 내 총 인증 횟수 합계
     @Query("SELECT COALESCE(SUM(r.verificationCount), 0) FROM RoundRecord r WHERE r.userChallenge.id = :userChallengeId")
-    Integer sumVerificationCountByUserChallengeId(@Param("userChallengeId") Long userChallengeId);
+    Long sumVerificationCountByUserChallengeId(@Param("userChallengeId") Long userChallengeId);
 }
